refactor(issue-34): uses Abstract for Error handling on Customer, Product and Cart controllers
